Refactor BlockBuildAlterEvent
Remove BaseBlockEvent class, this is only used by one class so no need
to have this extra class. Move all methods and fields. Remove the
deprecated setBuild method and update the test.
