+++ src/Event/Block/BlockBuildAlterEvent.php
@@ -2,12 +2,61 @@
 
 namespace Drupal\hook_event_dispatcher\Event\Block;
 
+use Drupal\Core\Block\BlockPluginInterface;
+use Drupal\hook_event_dispatcher\Event\EventInterface;
 use Drupal\hook_event_dispatcher\HookEventDispatcherInterface;
+use Symfony\Component\EventDispatcher\Event;
 
 /**
  * Class EntityInsertEvent.
  */
-class BlockBuildAlterEvent extends BaseBlockEvent {
+class BlockBuildAlterEvent extends Event implements EventInterface {
+
+  /**
+   * The build array.
+   *
+   * @var array
+   */
+  private $build;
+  /**
+   * The block.
+   *
+   * @var \Drupal\Core\Block\BlockPluginInterface
+   */
+  private $block;
+
+  /**
+   * BaseBlockEvent constructor.
+   *
+   * @param array $build
+   *   The build array.
+   * @param \Drupal\Core\Block\BlockPluginInterface $block
+   *   The block object.
+   */
+  public function __construct(array &$build, BlockPluginInterface $block) {
+    $this->build = &$build;
+    $this->block = $block;
+  }
+
+  /**
+   * Get the build array.
+   *
+   * @return array
+   *   The build array.
+   */
+  public function &getBuild() {
+    return $this->build;
+  }
+
+  /**
+   * Get the block.
+   *
+   * @return \Drupal\Core\Block\BlockPluginInterface
+   *   The block object.
+   */
+  public function getBlock() {
+    return $this->block;
+  }
 
   /**
    * {@inheritdoc}
